Seven Sages trade: cast the stored slot item to u8 explicitly
INV_CONTENT resolves to a u8 inventory slot and item->a is int16_t, so
the assignment was narrowing implicitly. MSVC flags that as C4244 and the
Windows CI build uses /WX, which would have failed the first Windows run
these commits have ever been through.

No behaviour change - the truncation already happened. The STORE_BOTTLE
case three lines below already casts the same field the same way.

+++ soh/soh/Enhancements/SevenSagesTrade/SevenSagesTrade.cpp
@@ -474,7 +474,7 @@ extern "C" bool SevenSagesTrade_GiveItem(int16_t randomizerGet) {
             return true;
 
         case STORE_SLOT:
-            INV_CONTENT(item->a) = item->a;
+            INV_CONTENT(item->a) = (u8)item->a;
             return true;
 
         case STORE_BOTTLE: {
